requirements.txt
uroman-python
tabulate
colorama
rich